feishu_rag/requirements.txt
chromadb==0.4.18
fastapi==0.104.1
tiktoken==0.5.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
sentence-transformers==2.2.2
torch==2.1.0
